fix: namechanges and moved files (sorry)
i'll probably squash some of these commits but i dont really understand git culture - i've been changing my mind about some minor things
