+++ Assets/Scripts/Data/InventoryEntry.cs.meta
