+++ Assets/Scripts/Engine/Inventory/InventoryUI.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2eee6b469bb7f466ab81cdc266ec42ab
