+++ Assets/Scripts/UI/InventoryItem.cs
@@ -11,29 +11,42 @@ namespace GoopGame.UI
     /// Uses ItemData scriptable object to display the correct images and text.
     /// </summary>
     public class InventoryItem : MonoBehaviour
-    {
-        //Hooking up UI elements in the editor so we can change them :3
-        [Header("UI references")]
+    {   
+        [Header("Components")]
         [SerializeField]
         private Image _image;                  //Reference to the image renderer
 
         [SerializeField]
         private TextMeshProUGUI _amountText;   //Reference to the text renderer
 
-        //TODO: I plan to have a description window as well :)
 
-        [Header("Public variables")]
+        [Header("UI-information (temp)")]
         public ItemData itemData;               //Reference to the ScriptableObject 'recipe' for item creation
-        public int amount = 1;
+
+        //IMPORTANT!! Almost none of these variables will be public later, they are only public now
+        //for testing purposes!
+        public string displayName;
+        public string displayDescription;
+        public int displayAmount = 1;
 
 
-        public void Start() //TODO: Change to Init()
+
+
+        public void Start() //TODO: Replace this entirely with the Init() function
         {
-            //TODO: Use Parameter to fetch the correct InventoryEntry
-            //itemData = entry.recipe;
-            //amount = entry.amount;
             UpdateUI();
         }
+
+        public void Init(InventoryEntry entry)
+        {
+            itemData = entry.Item;
+            displayName = itemData.Name;
+            displayDescription = itemData.Description;
+
+            displayAmount = entry.Amount;
+            UpdateUI();
+        }
+
 
         //Updates all of the visual elements in the InventoryItem gameobject to match ItemData recipe.
         public void UpdateUI()
@@ -41,8 +54,9 @@ namespace GoopGame.UI
             _image.sprite = itemData.Icon;
 
             //Set the amountText to current amount, or empty string if theres only 1 of the item
-            _amountText.text = amount > 1 ? amount.ToString() : string.Empty;
+            _amountText.text = displayAmount > 1 ? displayAmount.ToString() : string.Empty;
 
+            //TODO: Hover over item to see its name and description
         }
     }
 }
